NDRS-833: Move `Event` to a separate file.
